# Optimize `method_missing` with `alias_method`